Read instructions from the stream regardless
If the instructions are not read from the stream they might leak into the next instruction.
For example if a clipboard instruction is sent, but onclipboard is not set it will leak the clipboard content into the next instruction and make an uncaught exception if the next instructions would be an image and the leaked clipboard length is a multiple of 3. (the image base6 data to be drawn would be corrupted). I'm sure there might be other cases where this could cause uncaught exceptions.