GLEN-188: Backport toArray filter.
The changes backported for GLEN-169 are partially broken due to a
dependency on the toArray filter which was added upstream for
GUACAMOLE-526 as part of the development required to upgrade to
AngularJS 1.6.9.

The version of the toArray filter added with this commit is the current
upstream version, effectively isolated from the following commits:

 * f7ca11df5 - GUACAMOLE-526: Provide key/value pairs via toArray filter. Consistently return same array for same object.
 * 2ff980ded - GUACAMOLE-526: Remove debug console.log() from toArrayFilter().
 * b3eeb36b8 - GUACAMOLE-526: Update webapp to angular 1.6.9.

Other changes related to the upstream upgrade to AngularJS 1.6.9 remain
omitted for sake of API compatibility.